LayoutTests/http/tests/websocket/tests/hybi/url-attribute.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">

description(".url attribute should work, previously supported .URL should not.");

var url = "ws://127.0.0.1:8880/";
var ws = new WebSocket(url);

shouldBeEqualToString("ws.url", url);
shouldBeUndefined("ws.URL");

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
